backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.2",
    "fastembed>=0.8.0",
    "jinja2>=3.1.6",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langsmith>=0.7.22",
    "opencv-python>=4.13.0.92",
    "pdf2image>=1.17.0",
    "pgvector>=0.4.2",
    "playwright>=1.58.0",
    "psycopg[binary]>=3.3.3",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "scikit-image>=0.26.0",
    "uvicorn>=0.42.0",
]
